Add files via upload
This is a 3D printable case for the motherboard of the HP Elitebook 2540p. Note: you need to remove the motherboard from the laptop itself, this is a project for transforming this laptop into a thin desktop/home theater computer. It will require sticking the pieces together with super glue like i did, and it not very polished and/or easy to print/assemble/make, but i had fun making this project, althought i will not be improving it in the future.